Update build for i18n deploy workflow (#1743)
This PR updates the `build_and_deploy_i18n` workflow to run a simplified build before the acceptance tests instead of the full i18n build. This shortens the amount of time it takes before we see the results of the acceptance tests. The simplified build is the same as the `dev` build, but without the watch task. The i18n build runs in parallel and still must complete successfully before the deploy jobs are run.

The `build_and_deploy_hold` workflow also runs the i18n build, but another item (SLAP-2103) plans to remove the acceptance tests from this workflow, so it doesn't seem necessary to update it to run a simplified build.

J=SLAP-2099
TEST=auto

Tested on another branch (`test-basic-build`) and saw that the acceptance and unit tests run once the basic build is complete, the i18n build runs in parallel, and the deploy waits until both the tests and the i18n build finish.