Upgrade @percy/cli (#1751)
This PR adds back in the `@percy/cli` upgrade changes that were originally made in PR #1721 and reverted in #1741 due to failing Netlify checks. Since the Netlify checks are no longer needed, we can proceed with the upgrade. The version of `@percy/cli` is now the latest (v1.2.1) and, as noted in the original PR, node versions older than 14 will no longer be supported.

J=SLAP-2045, SLAP-2102
TEST=manual

Check that the Percy snapshots are generated correctly.