Update rtf-converter to v1.7.1 (#1783)
Update the version of the `rtf-converter` library to v1.7.1 to include a bug fix for an edge case where parsing bracketed text in `toTruncatedHTML` could throw an error.

J=TECHOPS-6710
TEST=manual

Use a local build of the SDK in the client's experience where this bug was seen. See that the entity that previously was throwing an error now displays successfully.