Modify `View All` URL based on Parent's URL.
This change mimics the `setParentUrl` in the `Navigation` Component. It's meant
to support the iFrame integration. In that integration clicking the `View All`
button (or openining it in a new tab) should mean you are still on the Parent's
domain.

T=TECHOPS-7751
TEST=manual

Tested manually in conjunction with the corresponding Theme change.