wcag: use appropriate aria roles for search input
A customer noticed that we had an aria-expanded attribute on the
search input, which they flagged as incorrect.  Based on
https://w3c.github.io/aria/#combobox, it appears we need to use
the combobox role on the input element and search on the form
element.

Additionally ran manual accessibility checks with Chrome Lighthouse
and Firefox to inspect the accessibility tree.  This revealed the
need to properly label the autocomplete results listbox control.

J=TECHOPS-8758
TEST=auto, manual

Reran acceptance tests