Update Github Actions to sync a separate EU build
Previously the GH actions were syncing the US
assets to GCP. This change updates the actions so
that a seprate EU build is ran, and only those assets
are synced to GCP. Also, remove the syncing of assets
from dev branches.

J=BACK-2268
TEST=none

I will test this with a test version (e.g. 0.0.10)
and confirm the build works as excpected