Create seprate EU artifact in github actions
Previously the US and EU builds were using the same artifact, but
were using separate directories. There was an issue with the
acceptance tests which I suspect was a bug with this approach, so
I am updating the flow to use separate artificts for the different
regions instead.

J=BACK-2268
TEST=none

I will create a test tag version and confirm the deploy works
correctly