Update Analytics for EU (#1827)
When the cloud region is EU, send analytics events to https://www.eu.yextevents.com

J=BACK-2362
TEST=manual

Build EU and US versions locally. Connected them to a test jambo site and confirmed the EU and US events fire properly. Checked the user_events table in the EU and confirmed the events are there.