Update Workflow Versions to Remove node16 Deprecation Warnings (#1877)
Notable Changes:

1. We added two new jobs: `merge_multiple_artifacts` and `overwrite_artifact` in `.github/workflows/build_i18n.yml` to handle merging multiple locale files into the `build-output-US` and `build-output-EU` artifacts. This was necessary since we have to handle no longer being able to have multiple jobs in the same workflow write to one artifact name and mutate it, [a breaking change](https://github.com/actions/upload-artifact/blob/main/docs/MIGRATION.md#multiple-uploads-to-the-same-named-artifact) in `action/upload-artifact@v4`. Since `build_and_deploy_i18n.yml` calls both `build.yml` and `build_i18n.yml`, we can't have both write an artifact called `build-output-US` unless we pass the `overwrite: true` input in the latter. The `action/upload-artifact/merge@v4` doesn't allow overwriting, so we need an additional step of merging and uploading with a temporary name and then downloading the merged artifact and re-uploading to overwrite the correct name.

2. We bumped `google-github-actions/auth` and `google-github-actions/upload-cloud-storage` from v0 to v2

3. Updated  `actions/checkout`, `actions/setup-node`, `actions/upload-artifact`, `actions/download-artifact`, and `aws-actions/configure-aws-credentials` actions to v4

[J=WAT-4020](https://yexttest.atlassian.net/browse/WAT-4020)
TEST=manual

Testing was done in a forked repo. View [actions](https://github.com/jesuyedavid/answers-search-ui/actions) to see testing details and test commits. We also checked the US/EU build artifact unzipped files to ensure they are uploaded to AWS/GCP properly and that they look similar to build files prior to this update. Finally, we did local testing from the Answers-Hitchhiker-Theme repo's test site that pointed to our dev assets in AWS. The `dev` bucket in GCP did not have the correct permissions to allow local testing, so this local testing will be done after merging this PR and the changes from develop are pushed to `canary` in GCP.