Answers SDK: add cloudChoice support
Now that cloudChoice is a param inside of search-core,
this CR adds support for setting it in the SDK, in the
init() function, including setting the old behavior as
the default. It is not needed elsewhere like with the
CLOUD_REGION setting, as that is used for things like
analytics tracking.

J=WAT-4375
TEST=auto

Ran test suite