Update custom card handling logic
We were missing some data being passed to the custom card when present
in the search-loading case. The generativeDirectAnswer property is what
contains the data passed to the custom card (as seen in addChild of
generativedirectanswercomponent.js). We now add the searchState to this
object, so that the custom card can have its own loading state logic.

J=WAT-4595
TEST=manual
Spun up test site with custom card in HH theme and saw the card behave
as expected in the loading state.