ksearch: add aria-label to submit buttons for the search bar
As can be seen in the referenced item, there are concerns
that our submit buttons having aria-hidden svgs can mess with
screen readers, even with the sr-only submitText. This PR
adds an explicit aria-label to the submit button.

J=WAT-5037
TEST=manual

Ran site locally, and in the Accessibility section, saw
the Aria attribute 'Submit' as expected